+++ poi-scratchpad/src/main/java/org/apache/poi/hemf/record/emfplus/HemfPlusObject.java
@@ -174,7 +174,7 @@ public class HemfPlusObject {
             EmfPlusObjectType objectType = getObjectType();
             assert (objectType != null);
 
-            int size = 0;
+            long size = 0;
 
             totalObjectSize = 0;
             int dataSize2 = (int) dataSize;
@@ -192,7 +192,7 @@ public class HemfPlusObject {
             objectData = objectType.constructor.get();
             size += objectData.init(leis, dataSize2, objectType, flags);
 
-            return size;
+            return Math.toIntExact(size);
         }
 
         @Override
